+++ app/[intent]/[city]/[beachSlug]/water-temp/page.tsx
@@ -6,7 +6,7 @@ import {
   generateBeachSubPageMetadata,
 } from "@/lib/utils/beach-sub-page-utils";
 
-export const dynamic = "force-dynamic";
+export const revalidate = 3600;
 
 interface PageProps {
   params: Promise<{